Make input and output files the same
It says that only when input is the same that output the error "Input and output file must differ" should be printed (as expected in that test cases). But here input and output are two different files, and program should work successfully in that case. And to make testcase num.5 working that input should be updated.